+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added microgrid grid actions to myems-api
 - added microgrid heatpump actions to myems-api
 - added microgrid load actions to myems-api
+- added microgrid photovoltaic actions to myems-api
 ### Changed
 
 ### Fixed